requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
attrs==24.3.0
ccxt==4.4.46
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cloudpickle==3.1.0
cryptography==44.0.0
dateparser==1.2.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
gym==0.26.2
gym-notices==0.0.8
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
pandas==2.2.3
propcache==0.2.1
pycares==4.5.0
pycparser==2.22
pycryptodome==3.21.0
python-binance==1.0.27
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
setuptools==75.7.0
six==1.17.0
sympy==1.13.1
ta==0.11.0
torch==2.5.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
websockets==14.1
yarl==1.18.3
